Authentication with firebase (#103)
# User Sign-Up Flow


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/1ee984ec-7786-436b-8ce0-8110c065df05

# Verifications for Sign-Up


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/23a68ba6-7425-416d-83cd-893fde82db9a

# Verifications for Log-In


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/2e1fae5b-775d-4e99-9cab-6fa1694082e5

# Credentials persisted between sessions and Biometric Sign-In


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/0ed66f5d-2c45-4f40-9db6-c2c128a66003

# Offline behavior


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/cf1639a7-1ac4-4018-9e46-19b09ee460e1

# User Sign-Out Flow


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/5e90785f-b6a4-4d55-8798-cff31da3072a

# Multi-Device support (iOS)


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/024cc464-992a-4dd3-9768-4c8e194f1cac

+++ unischedule/lib/repositories/events/events_repository.dart
@@ -3,6 +3,7 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:unischedule/models/models.dart';
 import 'package:unischedule/constants/constants.dart';
 import 'package:unischedule/services/services.dart';
+import 'package:unischedule/providers/providers.dart';
 
 part 'events_repository.g.dart';
 
@@ -27,10 +28,14 @@ class EventsRepositoryImpl extends EventsRepository {
 
   @override
   Future<List<EventModel>> fetchEvents() async {
+    final userId = ref.watch(authenticationStatusProvider)?.uid;
+    if (userId == null) {
+      throw Exception(StringConstants.unauthorizedRequest);
+    }
     List<EventModel> events = boxService.getAll();
     if (events.isEmpty) {
       try {
-        final response = await client.getRequest('user/0MebgXs8fBYREjDKMlwq/events');
+        final response = await client.getRequest('user/$userId/events');
         events = response.map<EventModel>((json) => EventModel.fromJson(json)).toList();
         Map<String, EventModel> eventMap = {
           for (var event in events) event.id: event
@@ -45,10 +50,14 @@ class EventsRepositoryImpl extends EventsRepository {
   }
 
   Future<void> syncLocalEvents() async{
+    final userId = ref.watch(authenticationStatusProvider)?.uid;
+    if (userId == null) {
+      throw Exception(StringConstants.unauthorizedRequest);
+    }
     var unsyncedEvents = eventSyncModelBox.getAll();
     for (var event in unsyncedEvents) {
           try {
-            final response = await client.postRequest('user/0MebgXs8fBYREjDKMlwq/events', data: event.toJson());
+            final response = await client.postRequest('user/$userId/events', data: event.toJson());
             boxService.put(response['eventId'], event);
             boxService.delete(event.id);
             eventSyncModelBox.delete(event.id);
@@ -59,21 +68,15 @@ class EventsRepositoryImpl extends EventsRepository {
         }
   }
 
-  
-
-  //Future<List<EventModel>> fetchEvents() async {
-  //  List<EventModel> events = await client.getRequest('user/0MebgXs8fBYREjDKMlwq/events') // TODO change endpoint
-  //    .then((response) => response.map<EventModel>((json) => EventModel.fromJson(json)).toList())
-  //    .catchError((error) => boxService.getAll());
-  //  return events;
-  //}
-
 @override
 Future<void> addEvent(EventModel event) async {
-
+  final userId = ref.watch(authenticationStatusProvider)?.uid;
+  if (userId == null) {
+    throw Exception(StringConstants.unauthorizedRequest);
+  }
   boxService.put(event.id, event);
    try {
-    final response = await client.postRequest('user/0MebgXs8fBYREjDKMlwq/events', data: event.toJson());
+    final response = await client.postRequest('user/$userId/events', data: event.toJson());
     boxService.put(response['eventId'], event);
     boxService.delete(event.id);
   } catch (error) {
